Aggiunto OIS; Fixato il linking di OpenGL; Cambiata Platform in una Singleton nativa alla piattaforma!

+++ src/Math.mm
@@ -7,7 +7,7 @@ using namespace Dojo;
 void Math::seedRandom( uint seed )
 {
 	if( !seed )
-		seed = time( NULL );
+		seed = (uint)time( NULL );
 	
 	srand( seed );
 }
